+++ apps/prairielearn/src/lib/code-caller/code-caller-native.js
@@ -1,14 +1,14 @@
 // @ts-check
-import * as _ from 'lodash';
+import _ from 'lodash';
 import * as path from 'node:path';
 import * as child_process from 'node:child_process';
 import { v4 as uuidv4 } from 'uuid';
 import debugfn from 'debug';
 
-import { FunctionMissingError } from './code-caller-shared';
+import { FunctionMissingError } from './code-caller-shared.js';
 import { logger } from '@prairielearn/logger';
-import { deferredPromise } from '../deferred';
-import { APP_ROOT_PATH, REPOSITORY_ROOT_PATH } from '../paths';
+import { deferredPromise } from '../deferred.js';
+import { APP_ROOT_PATH, REPOSITORY_ROOT_PATH } from '../paths.js';
 
 const debug = debugfn('prairielearn:code-caller-native');
 
@@ -29,7 +29,7 @@ const EXITED = Symbol('EXITED');
 
 /** @typedef {CREATED | WAITING | IN_CALL | RESTARTING | EXITING | EXITED} CodeCallerState */
 
-/** @typedef {import('./code-caller-native-types').CodeCallerNativeChildProcess} CodeCallerNativeChildProcess */
+/** @typedef {import('./code-caller-native-types.js').CodeCallerNativeChildProcess} CodeCallerNativeChildProcess */
 
 /**
  * @typedef {Object} ErrorData
@@ -73,8 +73,8 @@ const EXITED = Symbol('EXITED');
 
 */
 
-/** @typedef {import('./code-caller-shared').CodeCaller} CodeCaller */
-/** @typedef {import('./code-caller-shared').CallType} CallType */
+/** @typedef {import('./code-caller-shared.js').CodeCaller} CodeCaller */
+/** @typedef {import('./code-caller-shared.js').CallType} CallType */
 
 /**
  * @implements {CodeCaller}
@@ -145,7 +145,7 @@ export class CodeCallerNative {
   }
 
   /**
-   * @param {import('./code-caller-shared').PrepareForCourseOptions} options
+   * @param {import('./code-caller-shared.js').PrepareForCourseOptions} options
    */
   async prepareForCourse({ coursePath, forbiddenModules }) {
     this.debug('enter prepareForCourse()');
@@ -337,7 +337,7 @@ export class CodeCallerNative {
 
     /** @type {import('child_process').SpawnOptions} */
     const options = {
-      cwd: __dirname,
+      cwd: import.meta.dirname,
       // stdin, stdout, stderr, data, and restart confirmations
       stdio: ['pipe', 'pipe', 'pipe', 'pipe', 'pipe'],
       env,