+++ apps/prairielearn/src/pages/elementFiles/elementFiles.js
@@ -1,11 +1,11 @@
 // @ts-check
-const asyncHandler = require('express-async-handler');
+import asyncHandler from 'express-async-handler';
 import * as path from 'node:path';
 import { Router } from 'express';
 
-import * as chunks from '../../lib/chunks';
-import { config } from '../../lib/config';
-import { APP_ROOT_PATH } from '../../lib/paths';
+import * as chunks from '../../lib/chunks.js';
+import { config } from '../../lib/config.js';
+import { APP_ROOT_PATH } from '../../lib/paths.js';
 import { HttpStatusError } from '@prairielearn/error';
 
 const router = Router({ mergeParams: true });