+++ apps/prairielearn/src/pages/instructorInstanceAdminAccess/instructorInstanceAdminAccess.js
@@ -1,12 +1,12 @@
 // @ts-check
-const asyncHandler = require('express-async-handler');
+import asyncHandler from 'express-async-handler';
 import * as express from 'express';
 import { z } from 'zod';
 
 import * as sqldb from '@prairielearn/postgres';
 
 const router = express.Router();
-const sql = sqldb.loadSqlEquiv(__filename);
+const sql = sqldb.loadSqlEquiv(import.meta.url);
 
 router.get(
   '/',
@@ -22,7 +22,7 @@ router.get(
       }),
     );
 
-    res.render(__filename.replace(/\.js$/, '.ejs'), res.locals);
+    res.render(import.meta.filename.replace(/\.js$/, '.ejs'), res.locals);
   }),
 );