+++ apps/prairielearn/src/pages/navbarCourseSwitcher/navbarCourseSwitcher.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
-import asyncHandler = require('express-async-handler');
-import { NavbarCourseSwitcher } from './navbarCourseSwitcher.html';
-import { selectCoursesWithStaffAccess } from '../../models/course';
+import asyncHandler from 'express-async-handler';
+import { NavbarCourseSwitcher } from './navbarCourseSwitcher.html.js';
+import { selectCoursesWithStaffAccess } from '../../models/course.js';
 
 const router = Router();