+++ apps/prairielearn/src/tests/instructorAssessment.test.ts
@@ -1,11 +1,11 @@
 import { assert } from 'chai';
-import request = require('request');
+import request from 'request';
 import * as cheerio from 'cheerio';
-import * as _ from 'lodash';
+import _ from 'lodash';
 
-import * as helperServer from './helperServer';
-import * as helperQuestion from './helperQuestion';
-import * as helperExam from './helperExam';
+import * as helperServer from './helperServer.js';
+import * as helperQuestion from './helperQuestion.js';
+import * as helperExam from './helperExam.js';
 
 const locals: Record<string, any> = {};