+++ packages/session/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prairielearn/session
 
+## 3.0.0
+
+### Major Changes
+
+- 4f30b7e: Publish as native ESM
+
 ## 2.0.6
 
 ### Patch Changes