+++ packages/workspace-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prairielearn/workspace-utils
 
+## 1.0.21
+
+### Patch Changes
+
+- Updated dependencies [4f30b7e]
+  - @prairielearn/postgres@2.0.0
+
 ## 1.0.20
 
 ### Patch Changes